# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "GFW UMD Forest API",
        "description": "## Global Forest Watch - UMD Forest API\n\nIdentifies areas of tree cover loss\n\n| Property | Value |\n| --- | --- | --- |\n| Coverage | Global land area (excluding Antarctica and other Arctic islands) |\n| Resolution | 30 x 30 meters |\n| Source | Hansen/UMD/Google/USGS/NASA |\n| Timescale | 2000–2014 |\n| Units | Percents and hectares |\n| Updates | Annual |\n",
        "version": "1.0.0"
    },
    "host": "example.vizzuality.com",
    "schemes": [
        "https",
        "http"
    ],
    "produces": [
        "application/vnd.api+json"
    ],
    "paths": {
        "/umd-loss-gain": {
            "get": {
                "description": "a",
                "operationId": "world",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "UMD"
                ],
                "parameters": [
                    {
                        "name": "geostore",
                        "in": "query",
                        "description": "Hash of the geostore",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period of time (default '2001-01-01,2013-01-01')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "thresh",
                        "in": "query",
                        "description": "Threshold (default 30)",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Results obtained",
                        "schema": {
                            "$ref": "#/definitions/USEAPI"
                        }
                    },
                    "404": {
                        "description": "Geostore not found",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/umd-loss-gain/use/{name}/{id}": {
            "get": {
                "description": "a",
                "operationId": "use",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "UMD"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "description": "Analisys name (['mining', 'oilpalm', 'fiber', 'logging'])",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "cartodbid",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period of time (default '2001-01-01,2013-01-01')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "thresh",
                        "in": "query",
                        "description": "Threshold (default 30)",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Results obtained",
                        "schema": {
                            "$ref": "#/definitions/USEAPI"
                        }
                    },
                    "404": {
                        "description": "Geostore not found",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        },
        "/umd-loss-gain/wdpa/{id}": {
            "get": {
                "description": "a",
                "operationId": "wdpa",
                "produces": [
                    "application/vnd.api+json"
                ],
                "tags": [
                    "UMD"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "wdpa id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Period of time (default '2001-01-01,2013-01-01')",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "thresh",
                        "in": "query",
                        "description": "Threshold (default 30)",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Results obtained",
                        "schema": {
                            "$ref": "#/definitions/USEAPI"
                        }
                    },
                    "404": {
                        "description": "Geostore not found",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "500": {
                        "description": "Unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    },
                    "default": {
                        "description": "unexpected error",
                        "schema": {
                            "$ref": "#/definitions/Errors"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Errors": {
            "type": "object",
            "description": "Errors",
            "properties": {
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Error"
                    }
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int32",
                    "description": "A unique identifier for this particular occurrence of the problem."
                },
                "links": {
                    "type": "object",
                    "description": "A links object",
                    "properties": {
                        "about": {
                            "type": "string",
                            "description": "A link that leads to further details about this particular occurrence of the problem."
                        }
                    }
                },
                "status": {
                    "type": "string",
                    "description": "The HTTP status code applicable to this problem, expressed as a string value"
                },
                "code": {
                    "type": "string",
                    "description": "An application-specific error code, expressed as a string value"
                },
                "title": {
                    "type": "string",
                    "description": "A short, human-readable summary of the problem that SHOULD NOT change from occurrence to occurrence of the problem, except for purposes of localization."
                },
                "detail": {
                    "type": "string",
                    "description": "A human-readable explanation specific to this occurrence of the problem. Like title, this field's value can be localized"
                },
                "source": {
                    "type": "object",
                    "description": "An object containing references to the source of the error, optionally including any of the following members",
                    "properties": {
                        "pointer": {
                            "type": "string",
                            "description": "A JSON Pointer [RFC6901] to the associated entity in the request document"
                        },
                        "parameter": {
                            "type": "string",
                            "description": "A string indicating which URI query parameter caused the error."
                        }
                    }
                },
                "meta": {
                    "type": "object",
                    "description": "A meta object containing non-standard meta-information about the error."
                }
            }
        },
        "USEAPI": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/UMDBodyAPI"
                    }
                }
            }
        },
        "UMDBodyAPI": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string",
                    "description": "Resource type"
                },
                "id": {
                    "type": "string",
                    "description": "Resource id"
                },
                "attributes": {
                    "properties": {
                        "schema": {
                            "$ref": "#/definitions/UMDAPIModel"
                        }
                    }
                }
            }
        },
        "UMDAPIModel": {
            "type": "object",
            "properties": {
                "iso": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "thresh": {
                    "type": "integer"
                },
                "year": {
                    "type": "integer"
                },
                "id1": {
                    "type": "integer"
                },
                "extent": {
                    "type": "number",
                    "format": "float"
                },
                "extent_perc": {
                    "type": "number",
                    "format": "float"
                },
                "loss": {
                    "type": "number",
                    "format": "float"
                },
                "loss_perc": {
                    "type": "number",
                    "format": "float"
                },
                "gain": {
                    "type": "number",
                    "format": "float"
                },
                "total_gain": {
                    "type": "number",
                    "format": "float"
                },
                "gain_perc": {
                    "type": "number",
                    "format": "float"
                }
            }
        }
    }
}